+++ src/common/matrix/matrix.h
@@ -12,8 +12,8 @@ namespace Mino
 
 class Matrix;
 
-Matrix operator*(double const& coeff, Matrix rhs);
-Matrix operator*(Matrix lhs, double const& coeff);
+Matrix operator*(float const& coeff, Matrix rhs);
+Matrix operator*(Matrix lhs, float const& coeff);
 Matrix operator*(Matrix const& lhs, Matrix const& rhs);
 Matrix operator+(Matrix lhs, Matrix const& rhs);
 Matrix operator-(Matrix lhs, Matrix const& rhs);
@@ -23,49 +23,49 @@ class Matrix
 {
     int _columns = 0;
     int _rows = 0;
-    std::vector<double> values = {};
+    std::vector<float> values = {};
 
 public:
-    using iterator = std::vector<double>::iterator;
-    using const_iterator = std::vector<double>::const_iterator;
+    using iterator = std::vector<float>::iterator;
+    using const_iterator = std::vector<float>::const_iterator;
 
     class ColumnProxy
     {
     public:
-        ColumnProxy(int begin, int size, std::vector<double>& values)
+        ColumnProxy(int begin, int size, std::vector<float>& values)
             : begin(begin), _size(size), values(values)
         {
         }
         ColumnProxy(ColumnProxy const& other) : begin(other.begin), values(other.values) {}
         ~ColumnProxy() {}
 
-        double& operator[](int n)
+        float& operator[](int n)
         {
             if (n > _size) throw MatrixErrors::InvalidAccess(begin, n);
             return values[begin + n];
         }
-        operator std::vector<double>();
+        operator std::vector<float>();
         int size() { return _size; }
 
     private:
         int begin;
         int _size;
-        std::vector<double>& values;
+        std::vector<float>& values;
     };
 
     static Matrix transpose(Matrix const& matrix);
-    static Matrix uniform(int columns, int rows, double value = 1.0);
+    static Matrix uniform(int columns, int rows, float value = 1.0);
     static Matrix map(Matrix const& lhs, Matrix const& rhs,
-                      std::function<double(double, double)> callback);
+                      std::function<float(float, float)> callback);
     static Matrix hadamard(Matrix const& lhs, Matrix const& rhs);
     static Matrix dot(Matrix const& lhs, Matrix const& rhs) { return lhs * rhs; }
     static void assertEqualSize(Matrix const& lhs, Matrix const& rhs, std::string const& message);
-    static std::vector<double> invertValues(Matrix const& matrix);
+    static std::vector<float> invertValues(Matrix const& matrix);
 
     Matrix() = default;
     Matrix(int columns, int rows);
-    Matrix(std::vector<std::vector<double>> const& matrix);
-    Matrix(std::vector<double> const& matrix, int columns, int rows);
+    Matrix(std::vector<std::vector<float>> const& matrix);
+    Matrix(std::vector<float> const& matrix, int columns, int rows);
     Matrix(Matrix const& matrix) = default;
     Matrix(Matrix&& matrix) = default;
     ~Matrix() {}
@@ -73,22 +73,22 @@ public:
     ColumnProxy operator[](int n);
     Matrix& operator=(Matrix const& matrix) = default;
     Matrix& operator=(Matrix&& matrix) = default;
-    Matrix& operator*=(double const& coeff);
+    Matrix& operator*=(float const& coeff);
     Matrix& operator+=(Matrix const& matrix);
     Matrix& operator-=(Matrix const& matrix);
 
     bool operator==(Matrix const& rhs) const;
 
-    void set(int col, int row, double value);
-    double& get(int col, int row);
-    double const& get(int col, int row) const;
+    void set(int col, int row, float value);
+    float& get(int col, int row);
+    float const& get(int col, int row) const;
     ColumnProxy const getColumn(int col) const;
     Matrix transpose() const { return Matrix::transpose(*this); }
 
     int columns() const { return _columns; }
     int rows() const { return _rows; }
     int size() const { return _columns * _rows; }
-    double length() const;
+    float length() const;
     iterator begin() { return values.begin(); }
     const_iterator begin() const { return values.begin(); }
     iterator end() { return values.end(); }
